Refactor logging controls
